+++ lib/statsd.dart
@@ -1,3 +1,4 @@
+/// Statsd client library for Dart.
 library statsd;
 
 import 'dart:async';
@@ -8,4 +9,4 @@ import 'package:logging/logging.dart';
 part 'src/client.dart';
 part 'src/connection.dart';
 
-Logger _logger = new Logger('Statsd');
+Logger _logger = new Logger('statsd');