+++ CHANGELOG.md
@@ -4,7 +4,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
-- added logger
+## [0.1.1]
+
+- added logger (name 'statsd') as well as some logging for UDP connection.
+- added possibility to set prefix for metric names in StatsdClient.
+- fixed bug in UDP connection when it was binding to the same port as Statsd
+  server.
 
 ## [0.1.0]